refactor: split XMTPContext.tsx into focused modules
Splits the 1,366-line XMTPContext.tsx into 7 focused modules under
src/contexts/xmtp/:
- types.ts — all TypeScript interfaces, shared refs/state types, constants, helpers
- connection.ts — client initialization, auto-connect, wallet connect/disconnect
- conversations.ts — conversation list, sync, select, seed, refresh, remove
- messages.ts — send/receive messages, global stream management, reconnection
- members.ts — ZAO member discovery, DM creation, inbox resolution
- groups.ts — group create, leave, get members
- XMTPProvider.tsx — thin provider composing the hooks above
- index.ts — barrel export

The original XMTPContext.tsx is now a thin re-export file for backwards
compatibility. All existing imports continue to work unchanged.

Closes #38
