+++ src/contexts/xmtp/index.ts
@@ -1,0 +1,4 @@
+// Barrel export — re-exports everything from the split XMTP context modules.
+
+export { XMTPProvider, useXMTPContext, useXMTPContextSafe } from '@/contexts/xmtp/XMTPProvider';
+export type { ZaoMember, XMTPContextValue } from '@/contexts/xmtp/types';
